backend/requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.44.0
python-multipart==0.0.24
tensorflow==2.20.0
numpy==2.4.3
pillow==12.1.1
